refactor(ai): consolidate attachments into file package
